+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - PackageLookup#tools separates tool identifier and tool command.
 - PackageLookup#installed? checks for tool command before lookup.
+- Support for FreeBSD's pkgng added to PackageLookup.
 
 ## 0.0.1